+++ src/lib/index.ts
@@ -1,5 +1,3 @@
-import Line from "./line"
-import Point from "./point"
 import getArrow, { ArrowOptions } from "./getArrow"
 
-export { Line, Point, getArrow, ArrowOptions }
+export { getArrow, ArrowOptions }